text_to_video/requirements.txt
django==4.2.3
pillow==9.5.0
opencv-python==4.8.0.74
